pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pokemon-agents"
version = "0.1.0"
description = "AI agents that play Pokemon ROM games"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pyboy>=2.0.0",
    "gymnasium>=0.29.0",
    "pygame>=2.0.0",
    "numpy>=1.24.0",
    "Pillow>=10.0.0",
    "imageio[ffmpeg]>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
]
rl = [
    "stable-baselines3>=2.0.0",
    "sb3-contrib>=2.0.0",
    "torch>=2.0.0",
    "pyyaml>=6.0.0",
]
llm = [
    "anthropic>=0.25.0",
    "openai>=1.0.0",
    "boto3>=1.34.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
